createEnquetePageの完成
次のページの作成

+++ src/main/java/com/grouping/grouping_system/repository/IEnqueteRepository.java
@@ -19,5 +19,5 @@ public interface IEnqueteRepository {
      * @param enquete 追加する{@link Enquete}
      * @return 追加行数
      */
-    int insert(Enquete enquete);
+    long insert(Enquete enquete);
 }
